Add first pass of web interface
Adds a web interface that provides an indicator of server status, as well as buttons to manually turn the server on and off.

Adds the backing API required for the web interface functions - these return immediately and therefore the status indicator is immediately set to the desired state, which isn't true. I need to change this so the indicator is only set on a cadence of checking if the server is online, but that'l be later